requirements.txt
certifi==2022.6.15
charset-normalizer==2.0.12
idna==3.3
oauthlib==3.2.0
Pillow==8.4.0
python-dotenv==0.20.0
requests==2.27.1
requests-oauthlib==1.3.1
selenium==3.141.0
tweepy==4.6.0
urllib3==1.26.10
webdriver-manager==3.7.1
